requirements.txt
discord.py==2.3.1
Markdown==3.4.4
PyYAML==6.0.1
regex==2022.10.31